Rename csv-preview to yml-preview
